Remove upgrade instructions from the project (#18)
Updating dependencies is an just before you release is asking for problems.
Making a release is -assuming the process works- an activity wich needs a
fixed amount of time. Updating depencies is an unbounded time activity.

Additionally Renovate seems to be powerful enough to pickup all upgrades. If
some upgrades are not picked up, we should investigate renovate rather then
rely on a manual process.